moved SubGraph class into graph module.
