+++ NodeGraphQt/base/graph.py
@@ -13,7 +13,7 @@ from .commands import (NodeAddedCmd,
 from .factory import NodeFactory
 from .menu import NodeGraphMenu, NodesMenu
 from .model import NodeGraphModel
-from .node import NodeObject, BaseNode, SubGraph
+from .node import NodeObject, BaseNode
 from .port import Port
 from ..constants import (DRAG_DROP_ID,
                          PIPE_LAYOUT_CURVED,
@@ -1567,3 +1567,67 @@ class NodeGraph(QtCore.QObject):
             node (BaseNode): node object.
         """
         return self.get_node_by_id('0' * 13)
+
+
+class SubGraph(object):
+    """
+    The ``NodeGraphQt.SubGraph`` class is the base class that all
+    Sub Graph Node inherit from.
+
+    *Implemented on NodeGraphQt: * ``v0.1.0``
+
+    .. image:: _images/example_subgraph.gif
+        :width: 80%
+
+    """
+
+    def __init__(self):
+        self._children = set()
+
+    def children(self):
+        """
+        Returns the children of the sub graph.
+        """
+        return list(self._children)
+
+    def create_from_nodes(self, nodes):
+        """
+        Create sub graph from the nodes.
+
+        Args:
+            nodes (list[NodeGraphQt.NodeObject]): nodes to create the sub graph.
+        """
+        if self in nodes:
+            nodes.remove(self)
+        [n.set_parent(self) for n in nodes]
+
+    def add_child(self, node):
+        """
+        Add a node to the sub graph.
+
+        Args:
+            node (NodeGraphQt.BaseNode): node object.
+        """
+        self._children.add(node)
+
+    def remove_child(self, node):
+        """
+        Remove a node from the sub graph.
+
+        Args:
+            node (NodeGraphQt.BaseNode): node object.
+        """
+        if node in self._children:
+            self._children.remove(node)
+
+    def enter(self):
+        """
+        Action when enter the sub graph.
+        """
+        pass
+
+    def exit(self):
+        """
+        Action when exit the sub graph.
+        """
+        pass
